added db connections to azure

+++ BlocketLiteAPI/Program.cs
@@ -11,6 +11,7 @@ namespace BlocketLiteAPI
     {
         public static void Main(string[] args)
         {
+            Environment.SetEnvironmentVariable("ASPNETCORE_ENVIRONMENT", "Production");
             var host = CreateHostBuilder(args).Build();
 
             // migrate the database.  Best practice = in Main, using service scope
@@ -18,11 +19,22 @@ namespace BlocketLiteAPI
             {
                 try
                 {
-                    var context = scope.ServiceProvider.GetService<BlocketLiteContext>();
-                    // for demo purposes, delete the database & migrate on startup so 
-                    // we can start with a clean slate
-                    context.Database.EnsureDeleted();
-                    context.Database.EnsureCreated();
+                    var environment = (Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT"));
+                    if (environment == "Development")
+                    {
+                        var context = scope.ServiceProvider.GetService<BlocketLiteContext>();
+                        // for demo purposes, delete the database & migrate on startup so 
+                        // we can start with a clean slate
+                        context.Database.EnsureDeleted();
+                        context.Database.EnsureCreated();
+                    }
+                    if (environment == "Production")
+                    {
+                        var context = scope.ServiceProvider.GetService<BlocketLiteContext>();
+                        // For production we need to know that the DB is created
+                        context.Database.EnsureCreated();
+                    }
+
                 }
                 catch (Exception ex)
                 {
